refactor: move most datapage FAQ code over to `DatapageHelpers`
